move string functions to another file
